pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Django bookstore API"
authors = ["Fernando"]
packages = [
    {include = "bookstore"},
    {include = "product"},
    {include = "order"}
]

[tool.poetry.dependencies]
python = "^3.12"
Django = "5.2.1"
djangorestframework = "3.16.0"
django-cors-headers = "4.7.0"
django-debug-toolbar = "5.2.0"
django-extensions = "4.1"
psycopg2-binary = "^2.9.10"
dj-database-url = "^3.0.0"
python-decouple = "^3.8"
whitenoise = "^6.6.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-django = "4.11.1"
factory-boy = "3.3.3"
faker = "37.1.0"
black = "25.1.0"
flake8 = "^7.3.0"
coverage = "^7.11.3"
pytest-cov = "^5.0.0"
model-bakery = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
